refactor(pack): parse an image or buffer directive on its own record
The value of an image. or a bufferObject. line was read by two static
methods of ShaderProperties that touched nothing of the file's own state
and were the one reason it imported four of the value types. Each now
parses on the record it builds, ImageInformation.parse and
BufferObject.parse, beside the limit and the reading that were already
there; the properties file keeps the walk of the lines and the reasons
it drops one. Two imports of a type from its own package, left by the
move of the value types, go with it.
