+++ common/src/main/java/dev/vitrail/pack/model/BufferObject.java
@@ -1,6 +1,7 @@
 package dev.vitrail.pack.model;
 
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 
 /**
@@ -64,5 +65,66 @@ public record BufferObject(int index, long size, boolean relative, float scaleX,
 		return this.index
 				+ this.name.map(name -> " as " + name).orElse("")
 				+ " " + size;
+	}
+
+	/**
+	 * One {@code bufferObject.N} value, Iris's word counts, or a reason this line cannot be kept.
+	 * <p>
+	 * Returns null when the buffer was added. Two words or fewer are an absolute size and an
+	 * optional name; four or more are the relative form.
+	 */
+	public static String parse(String indexText, String value,
+			Map<Integer, BufferObject> buffers) {
+		int index;
+		try {
+			index = Integer.parseInt(indexText);
+		} catch (NumberFormatException e) {
+			return "index is not a number";
+		}
+
+		if (index > LIMIT) {
+			return "only indices 0 to " + LIMIT + " are allowed";
+		}
+
+		String[] parts = value.split(" ", -1);
+		if (parts.length == 0 || parts[0].isEmpty()) {
+			return "expected a size";
+		}
+
+		long size;
+		try {
+			size = Long.parseLong(parts[0]);
+		} catch (NumberFormatException e) {
+			return "size is not a number";
+		}
+
+		if (size < 1L) {
+			return "size below one disables the buffer";
+		}
+
+		if (parts.length <= 2) {
+			Optional<String> name = parts.length > 1 && !parts[1].isEmpty()
+					? Optional.of(parts[1])
+					: Optional.empty();
+			buffers.put(index, new BufferObject(index, size, false, 0.0F, 0.0F, name));
+			return null;
+		}
+
+		if (parts.length < 4) {
+			return "a relative buffer takes four words";
+		}
+
+		boolean relative = Boolean.parseBoolean(parts[1]);
+		float scaleX;
+		float scaleY;
+		try {
+			scaleX = Float.parseFloat(parts[2]);
+			scaleY = Float.parseFloat(parts[3]);
+		} catch (NumberFormatException e) {
+			return "relative scale is not a number";
+		}
+
+		buffers.put(index, new BufferObject(index, size, relative, scaleX, scaleY, Optional.empty()));
+		return null;
 	}
 }
